Create Browser Window Size website
This single page website will find the dimensions of the browser
viewport and will also let you copy the two dimensions by clicking on
their respective values.

The page has been designed to work with Internet Explorer 11 and the
latest versions of Chrome, Edge, Firefox, and Safari, as well as their
mobile variants.